Add files via upload
In this notebook we filter a dataset containing hourly values of f0F2 plus hourly values of the solar and geomagnetic parameters that influence electron content in the ionosphere; this filtering is done for a period of high solar activity. Periods of low and enhanced geomagnetic activity are identified for the enhanced solar activity dataset.